+++ Rabun.Oanda.Rest/Models/TransactionOrderCancel.cs
@@ -1,7 +1,8 @@
 ﻿namespace Rabun.Oanda.Rest.Models
 {
-public class TransactionOrderCancel: Transaction {
-    public int OrderId { get; set; }
-    public OandaTypes.Reason Reason { get; set; }
-}
+    public class TransactionOrderCancel : Transaction
+    {
+        public long OrderId { get; set; }
+        public OandaTypes.Reason Reason { get; set; }
+    }
 }
